requirements.txt
Flask==2.2.1
numpy==1.24.3
Pillow==9.3.0
tensorflow==2.15.0
pytest==8.3.3
pylint==2.15.0
pre-commit==2.20.0
pytest-cov==4.0.0
